feat: swtiched to floats

+++ Lazulite/Extensions.cs
@@ -15,11 +15,11 @@ public static class DisposableExtensions
 
 public static class MemoryBufferExtensions
 {
-    public static int AcceleratorIndex(this MemoryBuffer1D<double, Stride1D.Dense> buffer) => Compute.GetAcceleratorIndex(buffer.Accelerator);
-    public static AcceleratorStream GetStream(this MemoryBuffer1D<double, Stride1D.Dense> buffer) => Compute.GetStream(buffer.Accelerator);
-    public static void Return(this MemoryBuffer1D<double, Stride1D.Dense> buffer) => Compute.Return(buffer);
+    public static int AcceleratorIndex(this MemoryBuffer1D<float, Stride1D.Dense> buffer) => Compute.GetAcceleratorIndex(buffer.Accelerator);
+    public static AcceleratorStream GetStream(this MemoryBuffer1D<float, Stride1D.Dense> buffer) => Compute.GetStream(buffer.Accelerator);
+    public static void Return(this MemoryBuffer1D<float, Stride1D.Dense> buffer) => Compute.Return(buffer);
 
-    public static MemoryBuffer1D<double, Stride1D.Dense> DeferReturn(this MemoryBuffer1D<double, Stride1D.Dense> buffer)
+    public static MemoryBuffer1D<float, Stride1D.Dense> DeferReturn(this MemoryBuffer1D<float, Stride1D.Dense> buffer)
     {
         Compute.DeferReturn(buffer);
         return buffer;
@@ -28,7 +28,7 @@ public static class MemoryBufferExtensions
 
 public static class ArrayViewExtensions
 {
-    public static int AcceleratorIndex(this ArrayView1D<double, Stride1D.Dense> view) => Compute.GetAcceleratorIndex(view.GetAccelerator());
+    public static int AcceleratorIndex(this ArrayView1D<float, Stride1D.Dense> view) => Compute.GetAcceleratorIndex(view.GetAccelerator());
 }
 
 public static class AcceleratorExtensions
@@ -38,8 +38,8 @@ public static class AcceleratorExtensions
 
 public static class ValueExtensions
 {
-    public static ScalarValue AsScalar(this Value<double> value) => new(value.Data);
-    public static VectorValue AsVector(this Value<double[]> value) => new(value.Data);
-    public static MatrixValue AsMatrix(this Value<double[,]> value) => new(value.Data);
-    public static TensorValue3 AsTensorValue3(this Value<double[,,]> value) => new(value.Data);
+    public static ScalarValue AsScalar(this Value<float> value) => new(value.Data);
+    public static VectorValue AsVector(this Value<float[]> value) => new(value.Data);
+    public static MatrixValue AsMatrix(this Value<float[,]> value) => new(value.Data);
+    public static TensorValue3 AsTensorValue3(this Value<float[,,]> value) => new(value.Data);
 }
