+++ Lazulite/Values/ScalarValue.cs
@@ -4,15 +4,15 @@ using ILGPU.Runtime;
 
 namespace Lazulite.Values;
 
-public class ScalarValue : Value<double>
+public class ScalarValue : Value<float>
 {
-    public ScalarValue(double value, int aidx) : base(Compute.Get(aidx, 1)) => FromHost(value);
-    public ScalarValue(MemoryBuffer1D<double, Stride1D.Dense> buffer) : base(buffer) { }
+    public ScalarValue(float value, int aidx) : base(Compute.Get(aidx, 1)) => FromHost(value);
+    public ScalarValue(MemoryBuffer1D<float, Stride1D.Dense> buffer) : base(buffer) { }
     
-    public override double Unroll(double[] rolled) => rolled[0];
-    public override double[] Roll(double value) => [value];
+    public override float Unroll(float[] rolled) => rolled[0];
+    public override float[] Roll(float value) => [value];
     public override int[] GetShape(Index1D index) => [];
-    public override Value<double> Create(MemoryBuffer1D<double, Stride1D.Dense> buffer) => new ScalarValue(buffer);
+    public override Value<float> Create(MemoryBuffer1D<float, Stride1D.Dense> buffer) => new ScalarValue(buffer);
     
     public static ScalarValue operator +(ScalarValue a, ScalarValue b) => Compute.BinaryCall(Compute.ElementwiseAddKernels, a, b).AsScalar();
     public static ScalarValue operator -(ScalarValue a, ScalarValue b) => Compute.BinaryCall(Compute.ElementwiseSubtractKernels, a, b).AsScalar();
